Remove Faker support
With the recent happenings on Faker.js [1], I don't feel comfortable
including it out-the-box in test-data-bot anymore. This also helps us
solve #560, which was going to explore ways for the user to provide a
custom faker version. If someone does want to use Faker (or an
equivalent), they can now do that and make use of the `perBuild`
generator:

```js
const userBuilder = build('User', {
  fields: {
    name: perBuild(() => myFakeLibrary.randomName()),
    })
  }
```

And this can be wrapped by the caller if they desire:

```js
const fake = (callback) => {
  return perBuild(() => callback(myFakeLibrary))
}

const userBuilder = build('User', {
  fields: {
    name: fake => fake.randomName(),
  }
})
```

[1]: https://www.theverge.com/2022/1/9/22874949/developer-corrupts-open-source-libraries-projects-affected